Fix blind_spot_set being truncated through int8_t
UserLivoxLidarConfig::blind_spot_set is int32_t and the valid HAP
blind spot range is 50-200 cm, but the parsed value was cast through
int8_t: 200 became -56 (sent to the lidar as-is) and 255 became -1
(silently treated as "not configured"). Store the parsed int directly.

+++ src/parse_cfg_file/parse_livox_lidar_cfg.cpp
@@ -99,7 +99,7 @@ bool LivoxLidarConfigParser::ParseUserConfigs(const rapidjson::Document &doc,
       }
       user_config.blind_spot_set = -1;
     } else {
-      user_config.blind_spot_set = static_cast<int8_t>(config["blind_spot_set"].GetInt());
+      user_config.blind_spot_set = config["blind_spot_set"].GetInt();
     }
     if (!config.HasMember("dual_emit_en") || !config["dual_emit_en"].IsInt()) {
       if (config.HasMember("dual_emit_en")) {
